#  Compute transitive apps by id index lazily. 