Summary:
This is the first step to remove the app and pods index alltorgether.
For now it is caluclated only at one spot.

JIRA issues: MARATHON-7945