@jeschkies 2017-12-07
This actually does improve the performance a bit of the assign service ports benchmark.

## Master
```
[info] Benchmark                                        (numberOfGroups)  (numberOfSavedApps)  Mode  Cnt  Score   Error  Units
[info] GroupManagerBenchmark.assignDynamicServicePorts                 5                  100  avgt       0.091          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                 5                  500  avgt       0.521          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                 5                 1000  avgt       1.158          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                 5                 2500  avgt       3.330          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                 5                 5000  avgt       7.938          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                15                  100  avgt       0.084          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                15                  500  avgt       0.456          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                15                 1000  avgt       1.094          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                15                 2500  avgt       3.304          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                15                 5000  avgt       7.173          ms/op
```

## This Patch
```
[info] Benchmark                                        (numberOfGroups)  (numberOfSavedApps)  Mode  Cnt  Score   Error  Units
[info] GroupManagerBenchmark.assignDynamicServicePorts                 5                  100  avgt       0.056          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                 5                  500  avgt       0.299          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                 5                 1000  avgt       0.651          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                 5                 2500  avgt       2.187          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                 5                 5000  avgt       5.767          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                15                  100  avgt       0.065          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                15                  500  avgt       0.354          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                15                 1000  avgt       0.842          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                15                 2500  avgt       2.426          ms/op
[info] GroupManagerBenchmark.assignDynamicServicePorts                15                 5000  avgt       6.443          ms/op
```